requirements.txt
fairseq
librosa
matplotlib
numpy
PyYAML
resampy
scipy
SoundFile
tqdm
transformers
tensorboard